implement udpate_max_crossing_value_of_active_levels and switch from python's powell optimization to rust particle swarm

+++ crates/strategies/src/step/utils/entities/working_levels.rs
@@ -36,7 +36,7 @@ impl Default for BasicWLProperties {
     }
 }
 
-pub type WLMaxCrossingValue = f32;
+pub type WLMaxCrossingValue = Decimal;
 
 pub type WLIndex = u32;
